// src/app/core/interfaces/dasboard.interface.ts
export interface IDashboardData {
  address: object;
  age: number;
  bank: object;
  birthDate: string;
  bloodGroup: string;
  company: object;
  domain: string;
  ein: string;
  email: string;
  eyeColor: string;
  firstName: string;
  gender: string;
  hair: object;
  height: number;
  id: number;
  image: string;
  ip: string;
  lastName: string;
  macAddress: string;
  maidenName: string;
  password: string;
  phone: string;
  ssn: string;
  university: string;
  userAgent: string;
  username: string;
  weight: number;
}

export interface IMappingDashboardData {
  image: string;
  firstName: string;
  id: number;
  rotate: boolean;
}